Separated out frontend and backend data
I have a plan for how to order the ui and game in general,
I just need to actually implement it. The conplexity of control
between rendering the ui, taking in input, and simulating... its
a lot.

I just need to focus on one thing at a time. Implement any system
and accept that when I implement the next, I'm going to need to
revise all the others and that's okay.