Currently non-compiling for generics reasons
I don't think I know Rust well enough right now to program
in it the way I should be, so I'm going to take a break from
this project and work through the "The Rust Programming Language"
thing that comes with Rust.

The current issue is that I want a struct for pages of the UI, and
each one needs its own function to render itself, but that function
requires a frame with a generic type and I can't figure how to make
that work. I have a sneaking suspicion my design is fundamentally
flawed, so I'm hoping to find some best-practices type stuff in
the book.